requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
boto3==1.35.5
botocore==1.35.15
build==1.2.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.7
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
distro==1.9.0
emoji==2.12.1
et-xmlfile==1.1.0
fastapi==0.115.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.2
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.7
kubernetes==30.1.0
langchain==0.2.16
langchain-chroma==0.1.4
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langchain-upstage==0.1.8
langdetect==1.0.9
langsmith==0.1.107
layoutparser==0.3.4
libmagic==1.0
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openai==1.44.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.18.0
pikepdf==9.2.1
pillow==10.4.0
pinecone==5.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
portalocker==2.10.1
posthog==3.6.6
proto-plus==1.24.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pypandoc==1.13
pyparsing==3.1.4
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.1
pywin32==306
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scipy==1.14.1
selenium==4.23.1
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
starlette==0.38.5
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.9
unstructured-client==0.24.1
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.20.2

schedule~=1.2.2